\name{maigesActMod-class}
\docType{class}
\alias{maigesActMod-class}
\alias{maigesActMod}

\title{
  maigesActMod class, store results of functional classification of gene
  groups
}

\description{
  This class defines a structure to store the results of functional
  classification of gene groups (or modules) that was proposed by Segal
  et al. (2004), see the reference below. Objects of this type are
  generated by calling the function \code{\link{activeMod}}.
}

\section{Slots}{
  \describe{
    \item{\code{modBySamp}:}{numerical matrix containing the fraction of
      genes activated (or negative fraction of genes repressed) for all
      chips. The rows and columns of the matrix represents the chips and
      the gene groups used, respectively.}
    \item{\code{modByCond}:}{numerical matrix storing the fraction (or
      negative fraction) of different sample types that presents
      alteration in the gene groups tested. The rows and columns
      represents the sample types and gene groups, respectively.}
    \item{\code{globalScore}:}{list with the same length as the number
      of gene groups containing matrices with the genes as rows and 2
      columns. The first column gives a global score that measure the
      consistency of the classification of the gene and the group, the
      second column gives the p-value for this score, as described in
      Segal et al. (2004).}
    \item{\code{tissueScore}:}{a list similar to the previous one, but
      having arrays of 3 dimensions that gives scores similar to that
      one described above, but relating only to each specific sample
      type. Note that this new score was implemented in this package and
      not in the original Segal's work.}
    \item{\code{Date}:}{character string giving the date and time that
      the object was generated.}
    \item{\code{V.info}:}{list containg three characters. The first one is
      a string containing the R version used when the object was
      created. The second is a char vector with base packages and the
      last one is another char vector with additional packages and
      version numbers.}
  }
}

\details{
  Objects of this class are produced by calling the function
  \code{\link{activeMod}} over an object of class \code{\link{maiges}}
  to search for gene groups that present numbers of genes induced or
  repressed greater than the expected by chance in biological sample
  types of interest. This is done according to the model proposed by
  Segal et al. (2004), described below.
}

\section{Methods}{
  \describe{
    \item{image}{\code{signature(x = 'maigesActMod')}: image method for
      \code{\link{maigesActMod}} class. Display colour representation of
      the fractions of gene groups induced and repressed.}
    \item{plot}{\code{signature(x = 'maigesActMod')}: plot method for
      \code{\link{maigesActMod}} class. Do the same as image.}
    \item{print}{\code{signature(x = 'maigesActMod')}: print method for
      \code{\link{maigesActMod}} class.}
    \item{show}{\code{signature(x = 'maigesActMod')}: show method for
      \code{\link{maigesActMod}} class.}
    \item{summary}{\code{signature(x = 'maigesActMod')}: summary method for
      \code{\link{maigesActMod}} class.}
  }
}

\references{
  Segal, E.; Friedman, N.; Koller, D. and Regev, A. A module map showing
  conditional activity of expression modules in cancer. \bold{Nature
    Genetics}, 36, 1090-1098, 2004. (\url{http://www.nature.com/ng/journal/v36/n10/abs/ng1434.html})
}

\seealso{
  \code{\link{activeMod}}.
}

\author{
  Gustavo H. Esteves <\email{gesteves@vision.ime.usp.br}>
}

\keyword{classes}
