\name{maigesActNet-class}
\docType{class}
\alias{maigesActNet-class}
\alias{maigesActNet}

\title{
  maigesActNet class, store results of functional classification of gene
  networks
}

\description{
  This class defines a structure to store the results of functional
  classification of gene networks that was proposed and implemented
  in this package. Objects of this type are generated by calling the
  function \code{\link{activeNet}}.
}

\section{Slots}{
  \describe{
    \item{\code{scores}:}{numerical matrix storing the results of a
      statistic to test the functional activation of the networks
      studied. The rows and columns of the matrix represents the
      biological sample types and the networks, respectively.}
    \item{\code{Pvalues}:}{numerical matrix storing the p-values of the
      statistical test. As in the previous slot, matrix rows and columns 
      represents the biological sample types and the networks, respectively.}
    \item{\code{Date}:}{character string giving the date and time that
      the object was generated.}
    \item{\code{V.info}:}{list containg three characters. The first one is
      a string containing the R version used when the object was
      created. The second is a char vector with base packages and the
      last one is another char vector with additional packages and
      version numbers.}
  }
}

\details{
  Objects of this class are produced by calling the function
  \code{\link{activeNet}} over an object of class \code{\link{maiges}}
  to search for gene networks that present evidence of activation in different
  biological types. This is done according to the model proposed in the
  PhD thesis of the author of this package.
}

\section{Methods}{
  \describe{
    \item{image}{\code{signature(x = 'maigesActNet')}: image method for
      \code{\link{maigesActNet}} class. Display colour representation of
      the statistics or p-values of gene networks.}
    \item{plot}{\code{signature(x = 'maigesActNet')}: plot method for
      \code{\link{maigesActNet}} class. Do the same as image.}
    \item{print}{\code{signature(x = 'maigesActNet')}: print method for
      \code{\link{maigesActNet}} class.}
    \item{show}{\code{signature(x = 'maigesActNet')}: show method for
      \code{\link{maigesActNet}} class.}
    \item{summary}{\code{signature(x = 'maigesActNet')}: summary method for
      \code{\link{maigesActNet}} class.}
  }
}

\seealso{
  \code{\link{activeNet}}.
}

\author{
  Gustavo H. Esteves <\email{gesteves@vision.ime.usp.br}>
}

\keyword{classes}
